refactor: [jb-002] Refatoring Spring Security
- 안쓰는 import 제거
- UserDetails Entity로 이동
