+++ src/main/java/com/jjbacsa/jjbacsabackend/etc/config/SecurityConfig.java
@@ -3,7 +3,6 @@ package com.jjbacsa.jjbacsabackend.etc.config;
 import com.jjbacsa.jjbacsabackend.etc.filter.JwtTokenFilter;
 import com.jjbacsa.jjbacsabackend.etc.security.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
