+++ src/main/java/com/jjbacsa/jjbacsabackend/user/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.jjbacsa.jjbacsabackend.user.controller;
 
 import com.jjbacsa.jjbacsabackend.etc.dto.Token;
+import com.jjbacsa.jjbacsabackend.user.entity.CustomUserDetails;
 import com.jjbacsa.jjbacsabackend.user.dto.UserRequest;
 import com.jjbacsa.jjbacsabackend.user.dto.UserResponse;
 import com.jjbacsa.jjbacsabackend.user.service.UserService;
@@ -9,6 +10,7 @@ import io.swagger.annotations.Authorization;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
@@ -30,6 +32,9 @@ public class UserController {
     @ApiOperation( value = "",notes = "", authorizations = @Authorization(value = "Bearer +accessToken"))
     @GetMapping(value = "/user/me")
     public ResponseEntity<String> getMe() throws Exception{
+        System.out.println(((CustomUserDetails)SecurityContextHolder
+                .getContext().getAuthentication().getPrincipal())
+                .getUser().getAccount());
         //return new ResponseEntity<>(userService.getLoginUser(), HttpStatus.OK);
         return new ResponseEntity<>("success", HttpStatus.OK);
     }
